A lot of changes, most significant are: * [Tree] A tree can now have resizable column headers * [Tree] Fix the rendering of the first column of a tree in the LIST mode * [Scale] Implemented * [Scrollbar] Implemented * XCursor Support * Should fix the amd64 segv (I need a confirmation) * A lot of invisible things, bug fixes and mem leaks solved
* The canvas is broken


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@17792 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e17/libs/etk/src/lib/etk_signal.h
@@ -12,7 +12,7 @@
  */
 
 /** @brief Gets the offset of a member of a struct (mainly used for etk_signal_new) */
-#define ETK_MEMBER_OFFSET(type, member)   ((int)(&((type *)0)->member))
+#define ETK_MEMBER_OFFSET(type, member)   ((long)(&((type *)0)->member))
 
 /**
  * @struct Etk_Signal
